requirements.txt
discord.py
watchdog
mcrcon